Move tests fro CFamilyHelperTests to FileConfigTests
* no product or test changes
